cloud-functions/requirements.txt
# Required installs
google-api-python-client
oauth2client
google-cloud-pubsub
google-cloud-error-reporting
firebase-admin
pyyaml